# speed up logical indexing by bitarray