The recent PR [https://github.com/JuliaLang/julia/pull/29660](https://github.com/JuliaLang/julia/pull/29660) indicated that it is possible to significantly speed up logical indexing by bitarrays. So I consulted my Agner Fog, and wrote something faster.  Also, this will need benchmarking on a variety of architectures, since it depends on the `BLSR` and `TZCOUNT` instructions. The timings are on a 2 GHZ Broadwell, so multiply by two to get cycles / selected index.  

Setup:
```
julia> using BenchmarkTools, Random, Statistics
julia> function bench_getindex(d)
           for r in [0.01, 0.05, 0.1, 0.2, 0.5, 0.8, 0.9, 0.95, 0.99]
               m = d.<r 
               b = @benchmark(getindex($d,$m))
               bt = median(b.times)
               r = count(m)/length(d)
               println("logical indexing with $(r)=$(count(m))/$(length(d)) indices selected:")
               println("\t $(bt/1000) us = $(bt/count(m)) ns per selected index.\n")
           end
           nothing
       end;
julia> Random.seed!(23);
julia> d=rand(101*101);
```
Before:
```
julia> bench_getindex(d)
logical indexing with 0.008822664444662289=90/10201 indices selected:
	 10.213 us = 113.47777777777777 ns per selected index.

logical indexing with 0.04548573669248113=464/10201 indices selected:
	 14.399 us = 31.032327586206897 ns per selected index.

logical indexing with 0.09469659837270855=966/10201 indices selected:
	 18.122 us = 18.759834368530022 ns per selected index.

logical indexing with 0.19125575923929028=1951/10201 indices selected:
	 26.742 us = 13.706817016914403 ns per selected index.

logical indexing with 0.48514851485148514=4949/10201 indices selected:
	 59.125 us = 11.946857951101233 ns per selected index.

logical indexing with 0.7932555631800804=8092/10201 indices selected:
	 35.496 us = 4.38655462184874 ns per selected index.

logical indexing with 0.8954024115282816=9134/10201 indices selected:
	 28.356 us = 3.1044449310269324 ns per selected index.

logical indexing with 0.9454955396529752=9645/10201 indices selected:
	 25.62 us = 2.656298600311042 ns per selected index.

logical indexing with 0.9875502401725321=10074/10201 indices selected:
	 22.839 us = 2.267123287671233 ns per selected index.
```
after:
```
julia> bench_getindex(d)
logical indexing with 0.008822664444662289=90/10201 indices selected:
	 1.0089285714285714 us = 11.21031746031746 ns per selected index.

logical indexing with 0.04548573669248113=464/10201 indices selected:
	 2.0676 us = 4.4560344827586205 ns per selected index.

logical indexing with 0.09469659837270855=966/10201 indices selected:
	 2.9522222222222223 us = 3.056130664826317 ns per selected index.

logical indexing with 0.19125575923929028=1951/10201 indices selected:
	 5.131833333333333 us = 2.630360498889458 ns per selected index.

logical indexing with 0.48514851485148514=4949/10201 indices selected:
	 9.664 us = 1.952717720751667 ns per selected index.

logical indexing with 0.7932555631800804=8092/10201 indices selected:
	 13.934 us = 1.7219476025704399 ns per selected index.

logical indexing with 0.8954024115282816=9134/10201 indices selected:
	 15.2875 us = 1.673691701335669 ns per selected index.

logical indexing with 0.9454955396529752=9645/10201 indices selected:
	 16.04 us = 1.663037843442198 ns per selected index.

logical indexing with 0.9875502401725321=10074/10201 indices selected:
	 16.678 us = 1.6555489378598371 ns per selected index.
```

The main difference to my proposal in [https://github.com/JuliaLang/julia/pull/29660](https://github.com/JuliaLang/julia/pull/29660) is the following: We need one TZCNT per index, which has 3 cycles latency. We can avoid long dependency chains on TZCNT by using BLSR instead, with one cycle latency. Without this trick, this slows down significantly. I still don't get how this adds up to 3-5 cycles per index, but that's as fast as I could make it.

The main difference to the original code is that we replace hard to predict branches by dependency chains.