@chethega 2018-10-22
@mbauman Can you trigger a nanosoldier run?

I think the travis-ci fail is travis' problem, not this PR's problem. Unless there are change requests, I'm done with this PR, which supersedes [https://github.com/JuliaLang/julia/pull/29660](https://github.com/JuliaLang/julia/pull/29660). The `collect`->`findall` specialization is deliberately missing because the code here turns out to be significantly faster than `findall`. I don't know how to make `findall` fast without calling a slow linear->cartesian, so that will be a separate PR unless someone comes up with an idea.