@mbauman 2018-10-22
This is awesome, thank you!

@nanosoldier `runbenchmarks(ALL, vs = ":master")`