@nanosoldier 2018-10-23
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29746#issuecomment-431849617) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/06e148bbd6376a2b48db01b9fe50eab394a86e67/469527c_vs_c162219/report.md). cc @ararslan