@mbauman 2018-10-23
Hm, strange that this seems to show both time and memory regression in Nanosoldier.  They seem to be real:

```julia
julia> function perf_sumlogical(A)
           s = zero(eltype(A))
           nrows = size(A, 1)
           ncols = size(A, 2)
           r = falses(nrows)
           r[1:4:end] .= true
           @simd for i = 1:ncols
               val = @inbounds A[r, i]
               s += first(val)
           end
           return s
       end
perf_sumlogical (generic function with 1 method)

julia> using BenchmarkTools

julia> A = rand(100,100);

julia> @btime perf_sumlogical($A);
  24.910 μs (203 allocations: 31.44 KiB)

julia> @eval Base begin
       @inline _blsr(x)= x & (x-1)
       @inline function iterate(L::Base.LogicalIndex{Int,<:BitArray})
           L.sum == 0 && return nothing
           Bc = L.mask.chunks
           return iterate(L::Base.LogicalIndex{Int,<:BitArray}, (1, @inbounds Bc[1]))
       end
       @inline function iterate(L::Base.LogicalIndex{Int,<:BitArray}, s)
           Bc = L.mask.chunks
           i1, c = s
           while c==0
               i1 % UInt >= length(Bc) % UInt && return nothing
               i1 += 1
               @inbounds c = Bc[i1]
           end
           tz = trailing_zeros(c) + 1
           c = _blsr(c)
           return ((i1-1)<<6 + tz, (i1, c))
       end
       end
iterate (generic function with 201 methods)

julia> @btime perf_sumlogical($A);
  86.771 μs (603 allocations: 51.75 KiB)
```

The culprit seems to be the type assertion I've tagged below. Remove it and:

```julia
julia> @btime perf_sumlogical($A);
  14.063 μs (203 allocations: 31.44 KiB)
```