@nanosoldier 2018-10-24
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29746#issuecomment-432481878) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/6465e45b63fb8259f8e35492e06f6768f09df908/92bf1ca_vs_84024a1/report.md). cc @ararslan